projects/chapter-27-project/requirements.txt
# Chapter 27 Project Requirements

# Core AI library
openai>=1.0.0

# Vector database
chromadb>=0.4.0

# Document processing
pypdf>=3.17.0

# Environment variable management
python-dotenv>=1.0.0

# Token counting
tiktoken>=0.5.0
